Code cleanup. Separate serializers and viewsets. Added serializers and viewsets for  Notional Transfer Type
